+++ CHANGES.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- return one unified rule
 
 ## 0.1.3 - 2025-09-08
 ## 0.1.2 - 2025-09-08